calions_rustler v0.1.0 Calions.Rustler

Functions

gdpr_decrypt(a, b)

Decrypt an encrypted value with a key and return decrypted binary string

Returns:

{:ok, "decrypted_value"} in case of success

:error_decrypt when decrypting encrypted value with the given key failed

:error_value when parsing encrypted value failed

:error_key when parsing key failed

Examples

iex> Calions.Rustler.gdpr_decrypt("secret", "my_key")
{:ok, "secret"}

gdpr_encrypt(a)

Encrypt a binary string and return key with encrypted value

Returns:

{:ok, ["key", "encrypted_value"]} in case of success

:error_encrypt when generating key or encrypting value

Examples

iex> Calions.Rustler.gdpr_encrypt("secret")
{:ok, ["key", "encrypted_value"]}

gdpr_encrypt(a, b)

Encrypt a binary string with a key and return encrypted value

Returns:

{:ok, "encrypted_value"} in case of success

:error_encrypt_with_key when encrypting value with given key failed

:error_key when parsing key failed

Examples

iex> Calions.Rustler.gdpr_encrypt("secret", "my_key")
{:ok, "encrypted_value"}

gdpr_key()

Generate a GDPR key

Returns:

{:ok, "key"} in case of success

:error_key when generating key failed

Examples

iex> Calions.Rustler.gdpr_key()
{:ok, "my_key"}

ulid(a)

Generate a Monotonic ULID

Examples

iex> Calions.Rustler.ulid()
{:ok, "01EB7ETY5EH3VSY8GP945XATN2"}

uuid_v4(a)

Generate a V4 UUID

Examples

iex> Calions.Rustler.uuid_v4()
"42926399-b766-4b35-af5d-a71d820ceaf0"